update and rename the custom webclient, first attempt at decluttering the mainwindow class
