+++ TF2CompRosterChecker/CustomWebClient.cs
@@ -6,11 +6,11 @@ namespace TF2CompRosterChecker
 {
     //The default timeout for webrequests appears to be 100 seconds, 
     //which is far too much for our case, so we just make a new class
-    internal class TimeoutWebClient : WebClient
+    internal class CustomWebClient : WebClient
     {
         private readonly int timeout = 20 * 1000;
 
-        public TimeoutWebClient(int timeout) : base()
+        public CustomWebClient(int timeout) : base()
         {
             if (timeout > 0 && timeout < 20 * 1000)
             {
@@ -24,7 +24,7 @@ namespace TF2CompRosterChecker
             //Set a custom timeout
             w.Timeout = timeout; 
             //Set a user agent here, for privacy and interoperability.
-            w.UserAgent = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/110.0.0.0 Safari/537.36";
+            w.UserAgent = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/111.0.0.0 Safari/537.36";
             return w;
         }
     }
