+++ TF2CompRosterChecker/ETF2LChecker.cs
@@ -57,7 +57,7 @@ namespace TF2CompRosterChecker
                         List<Ban> bans = new List<Ban>();
 
 
-                        using (TimeoutWebClient wc = new TimeoutWebClient(8000))
+                        using (CustomWebClient wc = new CustomWebClient(8000))
                         {
                             wc.Encoding = Encoding.UTF8;
                             try
