+++ TF2CompRosterChecker/Utility.cs
@@ -1,0 +1,352 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Security.Cryptography;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Controls;
+using System.Windows.Documents;
+using System.Windows.Media;
+using System.Windows;
+using System.Text.RegularExpressions;
+
+namespace TF2CompRosterChecker
+{
+    internal class Utility
+    {
+        //Yes yes, ugli :<
+        public const string licenses = "Fody\n"
+            + "Copyright (c) Simon Cropp\n"
+            + "\n"
+            + "Permission is hereby granted, free of charge, to any person obtaining a copy \n"
+            + "of this software and associated documentation files (the \"Software\"), to deal \n"
+            + "in the Software without restriction, including without limitation the rights \n"
+            + "to use, copy, modify, merge, publish, distribute, sublicense, and/or sell \n"
+            + "copies of the Software, and to permit persons to whom the Software is \n"
+            + "furnished to do so, subject to the following conditions:\n"
+            + "\n"
+            + "The above copyright notice and this permission notice shall be included in \n"
+            + "all copies or substantial portions of the Software.\n"
+            + "\n"
+            + "THE SOFTWARE IS PROVIDED \"AS IS\", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR \n"
+            + "IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, \n"
+            + "FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE \n"
+            + "AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER \n"
+            + "LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING \n"
+            + "FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS \n"
+            + "IN THE SOFTWARE.\n"
+            + "\n"
+            + "-----------------------------------------------------------------------------\n"
+            + "\n"
+            + "Costura.Fody\n"
+            + "Copyright (c) Simon Cropp, Cameron MacFarland\n"
+            + "\n"
+            + "Permission is hereby granted, free of charge, to any person obtaining a copy \n"
+            + "of this software and associated documentation files (the \"Software\"), to deal \n"
+            + "in the Software without restriction, including without limitation the rights \n"
+            + "to use, copy, modify, merge, publish, distribute, sublicense, and/or sell \n"
+            + "copies of the Software, and to permit persons to whom the Software is \n"
+            + "furnished to do so, subject to the following conditions:\n"
+            + "\n"
+            + "The above copyright notice and this permission notice shall be included in \n"
+            + "all copies or substantial portions of the Software.\n"
+            + "\n"
+            + "THE SOFTWARE IS PROVIDED \"AS IS\", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR \n"
+            + "IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, \n"
+            + "FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE \n"
+            + "AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER \n"
+            + "LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING \n"
+            + "FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS \n"
+            + "IN THE SOFTWARE.\n"
+            + "\n"
+            + "-----------------------------------------------------------------------------\n"
+            + "\n"
+            + "Html Agility Pack (HAP)\n"
+            + "Copyright (c) ZZZ Projects, Simon Mourrier, Jeff Klawiter, Stephan Grell\n"
+            + "\n"
+            + "Permission is hereby granted, free of charge, to any person obtaining a copy \n"
+            + "of this software and associated documentation files (the \"Software\"), to deal \n"
+            + "in the Software without restriction, including without limitation the rights \n"
+            + "to use, copy, modify, merge, publish, distribute, sublicense, and/or sell \n"
+            + "copies of the Software, and to permit persons to whom the Software is \n"
+            + "furnished to do so, subject to the following conditions:\n"
+            + "\n"
+            + "The above copyright notice and this permission notice shall be included in \n"
+            + "all copies or substantial portions of the Software.\n"
+            + "\n"
+            + "THE SOFTWARE IS PROVIDED \"AS IS\", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR \n"
+            + "IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, \n"
+            + "FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE \n"
+            + "AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER \n"
+            + "LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING \n"
+            + "FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS \n"
+            + "IN THE SOFTWARE.\n"
+            + "\n"
+            + "-----------------------------------------------------------------------------\n"
+            + "\n"
+            + "Newtonsoft.Json\n"
+            + "Copyright (c) James Newton-King\n"
+            + "\n"
+            + "Permission is hereby granted, free of charge, to any person obtaining a copy \n"
+            + "of this software and associated documentation files (the \"Software\"), to deal \n"
+            + "in the Software without restriction, including without limitation the rights \n"
+            + "to use, copy, modify, merge, publish, distribute, sublicense, and/or sell \n"
+            + "copies of the Software, and to permit persons to whom the Software is \n"
+            + "furnished to do so, subject to the following conditions:\n"
+            + "\n"
+            + "The above copyright notice and this permission notice shall be included in \n"
+            + "all copies or substantial portions of the Software.\n"
+            + "\n"
+            + "THE SOFTWARE IS PROVIDED \"AS IS\", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR \n"
+            + "IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, \n"
+            + "FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE \n"
+            + "AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER \n"
+            + "LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING \n"
+            + "FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS \n"
+            + "IN THE SOFTWARE.\n"
+            + "\n"
+            + "-----------------------------------------------------------------------------\n"
+            + "\n"
+            + "AvalonEdit\n"
+            + "Copyright (c) AvalonEdit Contributors\n"
+            + "\n"
+            + "Permission is hereby granted, free of charge, to any person obtaining a copy\n"
+            + "of this software and associated documentation files (the \"Software\"), to deal\n"
+            + "in the Software without restriction, including without limitation the rights\n"
+            + "to use, copy, modify, merge, publish, distribute, sublicense, and/or sell\n"
+            + "copies of the Software, and to permit persons to whom the Software is\n"
+            + "furnished to do so, subject to the following conditions:\n"
+            + "\n"
+            + "The above copyright notice and this permission notice shall be included in all\n"
+            + "copies or substantial portions of the Software.\n"
+            + "\n"
+            + "THE SOFTWARE IS PROVIDED \"AS IS\", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR\n"
+            + "IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,\n"
+            + "FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE\n"
+            + "AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER\n"
+            + "LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,\n"
+            + "OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE\n"
+            + "SOFTWARE.\n"
+            + "-----------------------------------------------------------------------------\n"
+            + "\n"
+            + "TF2CompRosterChecker\n"
+            + "https://github.com/alekny/TF2CompRosterChecker\n"
+            + "\n"
+            + "MIT License\n"
+            + "\n"
+            + "Copyright (c) alekny 2023\n"
+            + "\n"
+            + "Permission is hereby granted, free of charge, to any person obtaining a copy\n"
+            + "of this software and associated documentation files (the \"Software\"), to deal\n"
+            + "in the Software without restriction, including without limitation the rights\n"
+            + "to use, copy, modify, merge, publish, distribute, sublicense, and/or sell\n"
+            + "copies of the Software, and to permit persons to whom the Software is\n"
+            + "furnished to do so, subject to the following conditions:\n"
+            + "\n"
+            + "The above copyright notice and this permission notice shall be included in all\n"
+            + "copies or substantial portions of the Software.\n"
+            + "\n"
+            + "THE SOFTWARE IS PROVIDED \"AS IS\", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR\n"
+            + "IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,\n"
+            + "FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE\n"
+            + "AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER\n"
+            + "LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,\n"
+            + "OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE\n"
+            + "SOFTWARE.\n";
+
+        public static string Sha256Hash(string input)
+        {
+            using (SHA256 sha256 = SHA256.Create())
+            {
+                byte[] inputBytes = System.Text.Encoding.UTF8.GetBytes(input);
+                byte[] hashBytes = sha256.ComputeHash(inputBytes);
+                string hash = BitConverter.ToString(hashBytes).Replace("-", "");
+                return hash;
+            }
+        }
+
+        //Functions for generating the table-content for our cases, with some overloadings.
+        public static TextBlock TBGen(string content, int marginleft, int margintop, bool bold)
+        {
+            TextBlock tb = new TextBlock
+            {
+                Text = content,
+                FontSize = 14,
+                HorizontalAlignment = HorizontalAlignment.Left,
+                VerticalAlignment = VerticalAlignment.Top,
+                TextWrapping = TextWrapping.NoWrap,
+                Height = 20
+            };
+            if (bold)
+            {
+                tb.FontWeight = FontWeights.Bold;
+            }
+            Thickness margin = tb.Margin;
+            margin.Left = marginleft;
+            margin.Top = margintop;
+            tb.Margin = margin;
+            return tb;
+        }
+
+        public static TextBlock TBGen(Run content, Run additional, Brush color, int marginleft, int margintop, bool bold)
+        {
+            TextBlock tb = new TextBlock();
+            tb.Inlines.Clear();
+            tb.Inlines.Add(content);
+            additional.Foreground = color;
+            tb.Inlines.Add(new Run(" "));
+            tb.Inlines.Add(additional);
+            tb.FontSize = 14;
+            tb.Height = 20;
+            if (bold)
+            {
+                tb.FontWeight = FontWeights.Bold;
+            }
+            tb.HorizontalAlignment = HorizontalAlignment.Left;
+            tb.VerticalAlignment = VerticalAlignment.Top;
+            tb.TextWrapping = TextWrapping.NoWrap;
+            Thickness margin = tb.Margin;
+            margin.Left = marginleft;
+            margin.Top = margintop;
+            tb.Margin = margin;
+            return tb;
+        }
+
+        public static TextBlock TBGen(Run content, Run additional, Brush color, Hyperlink link1, Hyperlink link2, int marginleft, int margintop, bool bold)
+        {
+            TextBlock tb = new TextBlock();
+            tb.Inlines.Clear();
+            tb.Inlines.Add(content);
+            additional.Foreground = color;
+            tb.Inlines.Add(new Run(" "));
+            tb.Inlines.Add(additional);
+            tb.Inlines.Add(link1);
+            tb.Inlines.Add(link2);
+            tb.FontSize = 14;
+            tb.Height = 20;
+            if (bold)
+            {
+                tb.FontWeight = FontWeights.Bold;
+            }
+            tb.HorizontalAlignment = HorizontalAlignment.Left;
+            tb.VerticalAlignment = VerticalAlignment.Top;
+            tb.TextWrapping = TextWrapping.NoWrap;
+            Thickness margin = tb.Margin;
+            margin.Left = marginleft;
+            margin.Top = margintop;
+            tb.Margin = margin;
+            return tb;
+        }
+
+        public static TextBlock TBGen(Hyperlink link, int marginleft, int margintop, bool bold)
+        {
+            TextBlock tb = new TextBlock();
+            tb.Inlines.Clear();
+            tb.Inlines.Add(link);
+            tb.FontSize = 14;
+            tb.Height = 20;
+            if (bold)
+            {
+                tb.FontWeight = FontWeights.Bold;
+            }
+            tb.HorizontalAlignment = HorizontalAlignment.Left;
+            tb.VerticalAlignment = VerticalAlignment.Top;
+            tb.TextWrapping = TextWrapping.NoWrap;
+            Thickness margin = tb.Margin;
+            margin.Left = marginleft;
+            margin.Top = margintop;
+            tb.Margin = margin;
+            return tb;
+        }
+
+        public static TextBlock TBGen(Run text, Hyperlink link, int marginleft, int margintop, bool bold)
+        {
+            TextBlock tb = new TextBlock();
+            tb.Inlines.Clear();
+            tb.Inlines.Add(text);
+            tb.Inlines.Add(new Run(" "));
+            tb.Inlines.Add(link);
+            tb.FontSize = 14;
+            tb.Height = 20;
+            if (bold)
+            {
+                tb.FontWeight = FontWeights.Bold;
+            }
+            tb.HorizontalAlignment = HorizontalAlignment.Left;
+            tb.VerticalAlignment = VerticalAlignment.Top;
+            tb.TextWrapping = TextWrapping.NoWrap;
+            Thickness margin = tb.Margin;
+            margin.Left = marginleft;
+            margin.Top = margintop;
+            tb.Margin = margin;
+            return tb;
+        }
+
+        public static TextBlock TBGen(Run text, Hyperlink link1, Hyperlink link2, int marginleft, int margintop, bool bold)
+        {
+            TextBlock tb = new TextBlock();
+            tb.Inlines.Clear();
+            tb.Inlines.Add(text);
+            tb.Inlines.Add(new Run(" "));
+            tb.Inlines.Add(link1);
+            tb.Inlines.Add(link2);
+            tb.FontSize = 14;
+            tb.Height = 20;
+            if (bold)
+            {
+                tb.FontWeight = FontWeights.Bold;
+            }
+            tb.HorizontalAlignment = HorizontalAlignment.Left;
+            tb.VerticalAlignment = VerticalAlignment.Top;
+            tb.TextWrapping = TextWrapping.NoWrap;
+            Thickness margin = tb.Margin;
+            margin.Left = marginleft;
+            margin.Top = margintop;
+            tb.Margin = margin;
+            return tb;
+        }
+        public static TextBlock TBGen(Run text, Hyperlink link1, Hyperlink link2, Hyperlink link3, int marginleft, int margintop, bool bold)
+        {
+            TextBlock tb = new TextBlock();
+            tb.Inlines.Clear();
+            tb.Inlines.Add(text);
+            tb.Inlines.Add(new Run(" "));
+            tb.Inlines.Add(link1);
+            tb.Inlines.Add(link2);
+            tb.Inlines.Add(link3);
+            tb.FontSize = 14;
+            tb.Height = 20;
+            if (bold)
+            {
+                tb.FontWeight = FontWeights.Bold;
+            }
+            tb.HorizontalAlignment = HorizontalAlignment.Left;
+            tb.VerticalAlignment = VerticalAlignment.Top;
+            tb.TextWrapping = TextWrapping.NoWrap;
+            Thickness margin = tb.Margin;
+            margin.Left = marginleft;
+            margin.Top = margintop;
+            tb.Margin = margin;
+            return tb;
+        }
+
+        public static int GetAllSteamIDs(string input)
+        {
+            string pattern = SteamIDTools.steamID3regex + "|" + SteamIDTools.profileUrlregex + "|"
+                + SteamIDTools.steamIDregex + "|" + SteamIDTools.profileCustomUrlregex + "|"
+                + SteamIDTools.etf2lProfileUrl + "|" + SteamIDTools.ugcProfileUrl + "|"
+                + SteamIDTools.rglProfileUrl + "|" + SteamIDTools.tf2centerProfileUrl;
+            MatchCollection matches = Regex.Matches(input, pattern);
+            return matches.Count;
+        }
+
+        //Thanks: https://stackoverflow.com/a/250400
+        public static DateTime UnixTimeStampToDateTime(string unixTimeStamp)
+        {
+            // Unix timestamp is seconds past epoch
+            DateTime dtDateTime = new DateTime(1970, 1, 1, 0, 0, 0, 0, DateTimeKind.Utc);
+            dtDateTime = dtDateTime.AddSeconds(Convert.ToDouble(unixTimeStamp)).ToLocalTime();
+            return dtDateTime;
+        }
+    }
+}
